Build EMT hash table by the LightSamplerBase
