+++ src/appleseed/renderer/kernel/lighting/backwardlightsampler.cpp
@@ -57,7 +57,6 @@ BackwardLightSampler::BackwardLightSampler(
     const Scene&                        scene,
     const ParamArray&                   params)
   : LightSamplerBase(params)
-  , m_emitting_triangle_hash_table(m_triangle_key_hasher)
 {
     // Read which sampling algorithm should the sampler use.
     m_use_light_tree = params.get_optional<string>("algorithm", "cdf") == "lighttree";
@@ -210,27 +209,6 @@ Dictionary BackwardLightSampler::get_params_metadata()
                             .insert("help", "Lights organized in a BVH"))));
 
     return metadata;
-}
-
-void BackwardLightSampler::build_emitting_triangle_hash_table()
-{
-    const size_t emitting_triangle_count = m_emitting_triangles.size();
-
-    m_emitting_triangle_hash_table.resize(
-        emitting_triangle_count > 0 ? next_pow2(emitting_triangle_count) : 0);
-
-    for (size_t i = 0; i < emitting_triangle_count; ++i)
-    {
-        const EmittingTriangle& emitting_triangle = m_emitting_triangles[i];
-
-        const EmittingTriangleKey emitting_triangle_key(
-            emitting_triangle.m_assembly_instance->get_uid(),
-            emitting_triangle.m_object_instance_index,
-            emitting_triangle.m_region_index,
-            emitting_triangle.m_triangle_index);
-
-        m_emitting_triangle_hash_table.insert(emitting_triangle_key, &emitting_triangle);
-    }
 }
 
 void BackwardLightSampler::sample_light_tree(
